Replace requirements.txt by minimal pyproject.toml
The line with kaleido had a parsing error (`=` instead of `==`) and the code
runs fine without it, so I didn't keep it. Similarly the git dependencies lead
to parsing errors because no module name was specified, since those also didn't
seem necessary to run the project I didn't keep them (I noticed that
https://github.com/lucyfarnik/sae-auto-interp depends on
"sae@git+https://github.com/EleutherAI/sae", but the project in that repository
has since been renamed to `sparsify`, so I don't think it would work as is).